feat(FN-3578): display line breaks in additional info
